+++ src/Exception/MiddlewareNotFound.php
@@ -1,0 +1,6 @@
+<?php
+namespace WoohooLabs\Harmony\Exception;
+
+class MiddlewareNotFound extends \Exception
+{
+}