+++ src/Exception/RouteNotFoundException.php
@@ -1,5 +1,5 @@
 <?php
-namespace WoohooLabs\Harmony\Router;
+namespace WoohooLabs\Harmony\Exception;
 
 class RouteNotFoundException extends \Exception
 {