+++ src/Middleware/MiddlewareInterface.php
@@ -1,17 +1,16 @@
 <?php
 namespace WoohooLabs\Harmony\Middleware;
 
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
 use WoohooLabs\Harmony\Harmony;
 
 interface MiddlewareInterface
 {
     /**
-     * @return string
+     * @param \Psr\Http\Message\ServerRequestInterface $request
+     * @param \Psr\Http\Message\ResponseInterface $response
+     * @param \WoohooLabs\Harmony\Harmony $next
      */
-    public function getId();
-
-    /**
-     * @param \WoohooLabs\Harmony\Harmony $harmony
-     */
-    public function execute(Harmony $harmony);
+    public function __invoke(ServerRequestInterface $request, ResponseInterface $response, Harmony $next);
 }